+++ src/main/java/com/gtnewhorizons/wdmla/api/ui/sizer/IPadding.java
@@ -27,11 +27,11 @@ public interface IPadding {
      */
     IPadding horizontal(int horizontal);
 
-    int getTop();
+    float getTop();
 
-    int getBottom();
+    float getBottom();
 
-    int getLeft();
+    float getLeft();
 
-    int getRight();
+    float getRight();
 }
